requirements.txt
tencentcloud-sdk-python-dnspod
requests